# perf(sharding): compressed morton code and gridpoints