@william-silversmith 2020-09-15
Do you want to break this into two PRs so we can get the first speedup merged? This is v. cool.